Integrate fundsTransfer and brokenPayments as webServices modules
