+++ .gitignore
@@ -62,6 +62,7 @@ typings/
 
 # dotenv environment variables file
 .env
+/modules/api/.env
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
