+++ modules/api/.env.schema
@@ -1,0 +1,7 @@
+WEB_PORT=
+WEB_SESSION_SECRET=
+
+SQL_DATABASE_HOST=
+SQL_DATABASE_USER=
+SQL_DATABASE_PASSWORD=
+SQL_DATABASE_DB_NAME=
