+++ modules/api/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@gateway/api",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "dev": "nodemon webServicesBackend.js",
+    "production": "pm2 start webServicesBackend.js"
+  },
+  "dependencies": {
+    "bitcoincashjs": "^0.1.14",
+    "dotenv": "^6.1.0",
+    "express-session": "^1.15.6",
+    "node-mysql": "^0.4.2",
+    "nodemon": "^1.18.7",
+    "pm2": "^3.2.2",
+    "session-file-store": "^1.2.0"
+  }
+}
